run/requirements.txt
Flask==3.0.3
Flask-MySQLdb==2.0.0
Werkzeug==3.0.2
